refacto: extract the proposal tally factory
